rename to manager and conflict check
