+++ config.example.yaml
@@ -8,7 +8,7 @@ mediaHost:
   # port of the media host (self) instance
   port: 444
   # connection information for media master
-master:
+manager:
   scheme: http
   host: "<SOME_HOST>"
   port: 443
