+++ internal/app/config.go
@@ -11,16 +11,16 @@ type SelfCfg struct {
 	Port   int    `yaml:"port"`
 }
 
-type masterCfg struct {
+type managerCfg struct {
 	Scheme string `yaml:"scheme"`
 	Host   string `yaml:"host"`
 	Port   int    `yaml:"port"`
 }
 
 type config struct {
-	Directories []string  `yaml:"directories"`
-	FileTypes   []string  `yaml:"fileTypes"`
-	Master      masterCfg `yaml:"master"`
+	Directories []string   `yaml:"directories"`
+	FileTypes   []string   `yaml:"fileTypes"`
+	Manager     managerCfg `yaml:"manager"`
 	SelfCfg     `yaml:"mediaHost"`
 }
