+++ internal/integration/manager/manager.go
@@ -1,5 +1,5 @@
-package master
+package manager
 
-type MasterApi interface {
+type ManagerApi interface {
 	RegisterNode(scheme string, host string, port int) error
 }
